refactor(web-scraping): error-notificationに倣って構造を整理
- domain層をinterfaces/とmodels/に分割
- 各レイヤーに__init__.pyを追加して明示的なエクスポート
- settings.pyをNoneチェックパターンに変更（テスト可能性向上）
- 循環importをTYPE_CHECKINGで解決
- git mvでファイル履歴を保持

変更内容:
- domain/interfaces/: インターフェース専用フォルダ作成
- domain/models/: モデル専用フォルダ作成
- config/__init__.py: 空ファイル作成
- presentation/__init__.py: main関数をエクスポート
- domain/__init__.py: error-notificationと同じパターンに更新

検証結果:
- テスト: 6/6 passed
- カバレッジ: 68.97% (要件60%クリア)
- Lint: All checks passed
