+++ lambda/web-scraping/src/domain/interfaces/dcp_scraper_interface.py
