+++ lambda/web-scraping/src/domain/interfaces/notifier_interface.py
@@ -1,13 +1,15 @@
 from abc import ABC, abstractmethod
+from typing import TYPE_CHECKING
 
-from domain import DcpAssets, DcpOpsIndicators
+if TYPE_CHECKING:
+    from domain.models import DcpAssets, DcpOpsIndicators
 
 
 class INotifier(ABC):
     """通知クラス(抽象クラス)"""
 
     @abstractmethod
-    def create_message_by(self, assets_info: DcpAssets, ops_indicators: DcpOpsIndicators) -> str:
+    def create_message_by(self, assets_info: "DcpAssets", ops_indicators: "DcpOpsIndicators") -> str:
         pass
 
     @abstractmethod
