+++ lambda/web-scraping/src/domain/interfaces/s3_repository_interface.py
