+++ lambda/web-scraping/src/domain/models/dcp_assets.py
