+++ lambda/web-scraping/src/domain/models/dcp_ops_indicators.py
