+++ lambda/web-scraping/src/domain/models/scraping_params.py
