+++ lambda/web-scraping/src/presentation/__init__.py
@@ -1,0 +1,7 @@
+"""Presentation レイヤー: Lambda イベント処理、依存性注入"""
+
+from .dcp_ops_notification import main
+
+__all__ = [
+    "main",
+]
